+++ cmd/people-svc/.env
@@ -1,0 +1,9 @@
+# Configuration Common to All Environments
+WIE_ENV=development
+DB_NAME=/private/var/ContasCorrentes/sqlite_dev.db
+PROXY_PORT=80
+GIN_PORT=8080
+FIBER_PORT=3000
+TMPL_ROOT=internal/templates/root_new
+PERSON_ROLES=System,Tenant,Application,Operator,Person 
+PERSON_STATUSES=Active,Inactive
