+++ internal/bff/views/home.html
@@ -1,0 +1,9 @@
+{{ define "content" }}
+<h1>Welcome</h1>
+
+<p>This is the BFF service.</p>
+
+<div style="margin-top:16px">
+  <a href="/people">Manage People</a>
+</div>
+{{ end }}
